+++ src/index.js
@@ -48,6 +48,8 @@ export { default as Toggle } from './Toggle/Toggle';
 export { default as Main } from './Main/Main';
 export { default as Card } from './Card/Card';
 export { default as ErrorBoundary } from './ErrorBoundary/ErrorBoundary';
+export { default as StateSelector } from './StateSelector/StateSelector';
+export { default as StartStopButton } from './StartStopButton/StartStopButton';
 
 export { default as SidePanel } from './SidePanel/SidePanel';
 export { Group, CollapsibleGroup } from './SidePanel/Group';
@@ -80,3 +82,5 @@ export { default as useHotKey } from './utils/useHotKey';
 export { currentPane, setCurrentPane } from './App/appLayout';
 
 export { getAppSpecificStore as getPersistentStore } from './utils/persistentStore';
+
+export { deviceInfo } from './Device/deviceInfo/deviceInfo';